Address PR remarks
